+++ admin/src/components/map/fairwayCardSetter.ts
